Fix KeyError in doctest from new xarray v0.14.0 release, close #455
Note, the previous method this commit is replacing does not
appear to be deprecated by the new xarray version. But, this change
is still needed to run properly, so this may be an upstream
regression in xarray.

My workaround was basically switching it from a by-value indexing to a boolean indexing operation. Either way, we take those indexed
values and plug them back into `res.coords[dim]`.